Compartimentalized functions of ExploreCircuitTable
