+++ src/components/explore-section/Circuit/ListView/ExploreCircuitTable.tsx
@@ -1,13 +1,11 @@
 'use client';
 
-import { InfoCircleOutlined } from '@ant-design/icons';
-import { Table, Tooltip } from 'antd';
+import { Table } from 'antd';
 import { ColumnsType, TableRowSelection } from 'antd/es/table/interface';
-import React, { Key, useState } from 'react';
-import { Resizable, ResizeCallbackData } from 'react-resizable';
+import React, { Key, useCallback, useMemo, useState } from 'react';
+import { ResizeCallbackData } from 'react-resizable';
 
 import CIRCUITS_FULL from '../content/circuits_tree_formatted';
-import { ArrowSmall } from '../icon/ArrowSubcircuitIcon';
 import { CircuitSchemaProps } from '../type';
 import getExpandableRowKeys from '../utils/getExpandableRowKey';
 
@@ -20,51 +18,12 @@ import NumericFilters from './NumericFilters';
 import columns from './Columns';
 import SearchBar from './SearchBar';
 
+import DownloadCircuitButton from './DownloadCircuitButton';
+import ResizableTitle from './ResizableTitle';
+import SubcircuitTable from './SubcircuitsTable';
+
 import { classNames } from '@/util/utils';
-
 import styles from './ExploreCircuiteTable.module.scss';
-
-interface ResizableTitleProps {
-  onResize?: (e: React.SyntheticEvent, data: ResizeCallbackData) => void;
-  width?: number;
-  [key: string]: any;
-}
-
-export function ResizableTitle(props: ResizableTitleProps) {
-  const { onResize, width, ...restProps } = props;
-
-  if (!width) {
-    // eslint-disable-next-line react/jsx-props-no-spreading
-    return <th {...restProps} />;
-  }
-
-  const handleKeyDown = (e: React.KeyboardEvent<HTMLSpanElement>) => {
-    if (e.key === 'Enter' || e.key === ' ') {
-      e.stopPropagation();
-    }
-  };
-
-  return (
-    <Resizable
-      width={width}
-      height={0}
-      handle={
-        <span
-          className="resize-handle"
-          role="button"
-          tabIndex={0}
-          aria-label={`Resize ${restProps.title || 'column'}`}
-          onClick={(e) => e.stopPropagation()}
-          onKeyDown={handleKeyDown}
-        />
-      }
-      onResize={onResize}
-    >
-      {/* eslint-disable-next-line react/jsx-props-no-spreading */}
-      <th {...restProps} />
-    </Resizable>
-  );
-}
 
 export default function ExploreCircuitTable() {
   // STATES
@@ -89,27 +48,35 @@ export default function ExploreCircuitTable() {
   const [minValue, setMinValue] = useState<number | undefined>(undefined);
   const [maxValue, setMaxValue] = useState<number | undefined>(undefined);
 
-  const handleResize =
+  const handleResize = useCallback(
     (key: string) =>
-    (e: React.SyntheticEvent, { size }: ResizeCallbackData) => {
-      setColumnWidths((prev) => ({
-        ...prev,
-        [key]: size.width,
-      }));
-    };
+      (e: React.SyntheticEvent, { size }: ResizeCallbackData) => {
+        setColumnWidths((prev) => ({
+          ...prev,
+          [key]: size.width,
+        }));
+      },
+    []
+  );
 
-  const handleExpandRow = (row: CircuitSchemaProps, _index: number) => {
-    if (!row.hasSubcircuits) return;
+  const handleExpandRow = useCallback((expanded: boolean, row: CircuitSchemaProps) => {
+    const rowKey = row.key;
+    setExpandedRowKeys((prev) =>
+      expanded ? [...prev, rowKey] : prev.filter((key) => key !== rowKey)
+    );
+  }, []);
+
+  const handleRowExpandClick = useCallback((row: CircuitSchemaProps, _index: number) => {
     const rowKey = row.key;
     setExpandedRowKeys((prev) =>
       prev.includes(rowKey) ? prev.filter((key) => key !== rowKey) : [...prev, rowKey]
     );
-  };
+  }, []);
 
   const mergedColumns: ColumnsType<CircuitSchemaProps> = columns(
     expandedRowKeys,
     calculateSubcircuitsForParent,
-    handleExpandRow,
+    handleRowExpandClick,
     handleResize
   ).map((col) => ({
     ...col,
@@ -126,65 +93,18 @@ export default function ExploreCircuitTable() {
     });
   }
 
-  const rowSelection: TableRowSelection<CircuitSchemaProps> = {
-    type: 'radio',
-    selectedRowKeys,
-    onChange: (newSelectedRowKeys: Key[], _selectedRows: CircuitSchemaProps[]) => {
-      const key = newSelectedRowKeys[0] as string;
-
-      const updatedKeys = selectedRowKeys[0] === key ? [] : [key];
-      setSelectedRowKeys(updatedKeys);
-    },
-  };
-
-  const expandedRowRender = (circuit: CircuitSchemaProps): JSX.Element => {
-    return (
-      <div className="relative flex flex-col">
-        <div className="relative flex flex-row pl-2">
-          <ArrowSmall iconColor="#8C8C8C" className="relative -top-0.5" />
-          <span className="ml-3 pb-2 text-base font-semibold uppercase tracking-wider text-[#8C8C8C]">
-            Subcircuits
-          </span>
-        </div>
-        <Table<CircuitSchemaProps>
-          className={classNames(
-            '[&_.ant-table-tbody]:bg-[#FAFAFA]',
-            '[&_.ant-table-row]:bg-[#FAFAFA]',
-            '[&_.ant-table-thead_th]:!text-sm',
-            '[&_.ant-table-thead_th]:!font-normal',
-            '[&_.ant-table-thead_th]:!text-[#8C8C8C]',
-            '[&_.ant-table-thead_th]:uppercase',
-            '[&_.ant-table-thead_th]:tracking-[0.05em]',
-            '[&_.ant-table-tbody > tr:last-child > td]:border-b-0',
-            '[&_.ant-table-thead > tr > th]:border-b-0',
-            '[&_.ant-table-expand-icon-col]:w-0',
-            '[&_.ant-table-expand-icon-col]:hidden',
-            styles.circuitTable
-          )}
-          components={{
-            header: {
-              cell: ResizableTitle,
-            },
-          }}
-          columns={mergedColumns}
-          dataSource={circuit.subcircuit || []}
-          pagination={false}
-          rowSelection={rowSelection}
-          expandable={{
-            expandedRowRender,
-            expandedRowKeys,
-            onExpand: (expanded: boolean, row: CircuitSchemaProps) => {
-              const rowKey = row.key;
-              setExpandedRowKeys((prev) =>
-                expanded ? [...prev, rowKey] : prev.filter((key) => key !== rowKey)
-              );
-            },
-            expandIcon: () => null,
-          }}
-        />
-      </div>
-    );
-  };
+  const rowSelection = useMemo(
+    (): TableRowSelection<CircuitSchemaProps> => ({
+      type: 'radio',
+      selectedRowKeys,
+      onChange: (newSelectedRowKeys: Key[], _selectedRows: CircuitSchemaProps[]) => {
+        const key = newSelectedRowKeys[0] as string;
+        const updatedKeys = selectedRowKeys[0] === key ? [] : [key];
+        setSelectedRowKeys(updatedKeys);
+      },
+    }),
+    [selectedRowKeys]
+  );
 
   const selectedRows = CIRCUITS_FULL.flatMap((circuit) =>
     circuit.key === selectedRowKeys[0]
@@ -194,6 +114,19 @@ export default function ExploreCircuitTable() {
 
   const lastRow = selectedRows.at(-1);
   const fileUrl = lastRow?.files?.[0]?.url;
+
+  const renderSubcircuits = useCallback(
+    (circuit: CircuitSchemaProps) => (
+      <SubcircuitTable
+        circuit={circuit}
+        mergedColumns={mergedColumns}
+        rowSelection={rowSelection}
+        expandedRowKeys={expandedRowKeys}
+        onExpand={handleExpandRow}
+      />
+    ),
+    [mergedColumns, rowSelection, expandedRowKeys, handleExpandRow]
+  );
 
   return (
     <div className="relative flex w-full flex-col pt-10">
@@ -234,7 +167,7 @@ export default function ExploreCircuitTable() {
         pagination={false}
         rowSelection={rowSelection}
         expandable={{
-          expandedRowRender,
+          expandedRowRender: renderSubcircuits,
           expandedRowKeys,
           onExpand: (expanded: boolean, row: CircuitSchemaProps) => {
             const rowKey = row.key;
@@ -246,31 +179,7 @@ export default function ExploreCircuitTable() {
         }}
       />
 
-      {fileUrl && (
-        <a
-          href={fileUrl}
-          type="button"
-          className="absolute bottom-6 right-10 flex h-20 w-[150px] items-center justify-center bg-primary-8 text-xl transition-bottom duration-300 ease-in-out"
-          style={{
-            visibility: selectedRowKeys && selectedRowKeys.length > 0 ? 'visible' : 'hidden',
-          }}
-        >
-          <span>Download</span>
-          <Tooltip
-            title={
-              <a
-                href="https://github.com/openbraininstitute/ConnectomeUtilities/blob/main/README.md"
-                target="_blank"
-                onClick={(e) => e.stopPropagation()}
-              >
-                The connectome will be downloaded in Connectome Utilities format, see more here.
-              </a>
-            }
-          >
-            <InfoCircleOutlined className="ml-2" />
-          </Tooltip>
-        </a>
-      )}
+      {fileUrl && <DownloadCircuitButton fileUrl={fileUrl} selectedRowKeys={selectedRowKeys} />}
     </div>
   );
 }
