+++ src/components/explore-section/Circuit/ListView/ResizableTitle.tsx
@@ -1,0 +1,43 @@
+import { Resizable, ResizeCallbackData } from 'react-resizable';
+
+export type ResizableTitleProps = {
+  onResize?: (e: React.SyntheticEvent, data: ResizeCallbackData) => void;
+  width?: number;
+  [key: string]: any;
+};
+
+export default function ResizableTitle(props: ResizableTitleProps) {
+  const { onResize, width, ...restProps } = props;
+
+  if (!width) {
+    // eslint-disable-next-line react/jsx-props-no-spreading
+    return <th {...restProps} />;
+  }
+
+  const handleKeyDown = (e: React.KeyboardEvent<HTMLSpanElement>) => {
+    if (e.key === 'Enter' || e.key === ' ') {
+      e.stopPropagation();
+    }
+  };
+
+  return (
+    <Resizable
+      width={width}
+      height={0}
+      handle={
+        <span
+          className="resize-handle"
+          role="button"
+          tabIndex={0}
+          aria-label={`Resize ${restProps.title || 'column'}`}
+          onClick={(e) => e.stopPropagation()}
+          onKeyDown={handleKeyDown}
+        />
+      }
+      onResize={onResize}
+    >
+      {/* eslint-disable-next-line react/jsx-props-no-spreading */}
+      <th {...restProps} />
+    </Resizable>
+  );
+}
